+++ components/Container/Container.module.css
@@ -32,10 +32,17 @@
   position: relative;
   width: 2.5rem;
   height: 2.5rem;
+  margin-left: 1.5rem;
   border-radius: 50%;
   background-color: grey;
   cursor: pointer;
   overflow: hidden;
+}
+
+.header_profile_theme_wrapper {
+  display: flex;
+  justify-content: center;
+  align-items: center;
 }
 
 .main {
