+++ components/Container/ThemeButton/ThemeButton.module.css
@@ -1,0 +1,8 @@
+.button {
+  position: relative;
+  width: 1.5rem;
+  height: 1.5rem;
+  background: none;
+  border: none;
+  cursor: pointer;
+}
