+++ components/Container/index.tsx
@@ -1,16 +1,11 @@
 import type { ReactNode } from "react";
-import { useState, useEffect } from "react";
 import Head from "next/head";
 import Image from "next/image";
 import { useRouter } from "next/router";
 
-import {
-  getTheme,
-  setTheme as setDocumentTheme,
-} from "../../common/utils/getTheme";
-
 import Nav from "../Nav";
 import Footer from "../Footer";
+import ThemeButton from "./ThemeButton";
 
 import styles from "./Container.module.css";
 
@@ -19,28 +14,11 @@ interface ContainerProps {
 }
 
 const Container = (props: ContainerProps) => {
-  const [theme, setTheme] = useState<string | null>(null);
   const router = useRouter();
 
   const handleProfileClick = () => {
     router.push("/");
   };
-
-  const handleChangeTheme = () => {
-    if (theme === "dark") {
-      setTheme("light");
-      setDocumentTheme("light");
-    }
-
-    if (theme === "light") {
-      setTheme("dark");
-      setDocumentTheme("dark");
-    }
-  };
-
-  useEffect(() => {
-    setTheme(getTheme());
-  }, []);
 
   return (
     <div className={styles.wrapper}>
@@ -50,11 +28,11 @@ const Container = (props: ContainerProps) => {
       <div className={styles.content_wrapper}>
         <header className={styles.header}>
           <Nav />
-          {(theme === "dark" || theme === "light") && (
-            <button onClick={handleChangeTheme}>버튼 {theme}</button>
-          )}
-          <div onClick={handleProfileClick} className={styles.header_profile}>
-            <Image src="/images/profile.jpg" alt="profile" layout="fill" />
+          <div className={styles.header_profile_theme_wrapper}>
+            <ThemeButton />
+            <div onClick={handleProfileClick} className={styles.header_profile}>
+              <Image src="/images/profile.jpg" alt="profile" layout="fill" />
+            </div>
           </div>
         </header>
         <main className={styles.main}>{props.children}</main>
